feat(webapp): refactor Home component and update layout structure
