+++ webapp/src/app/layout.tsx
@@ -1,7 +1,8 @@
-import type { Metadata } from 'next';
+'use client';
+
 import { Geist, Geist_Mono } from 'next/font/google';
 import './globals.css';
-import { GlobalLayout } from './_components/GlobalLayout';
+import dynamic from 'next/dynamic';
 
 const geistSans = Geist({
   variable: '--font-geist-sans',
@@ -13,10 +14,10 @@ const geistMono = Geist_Mono({
   subsets: ['latin'],
 });
 
-export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-};
+const GlobalLayout = dynamic(
+  () => import('@/components/GlobalLayout').then((mod) => mod.GlobalLayout),
+  { ssr: false },
+);
 
 export default function RootLayout({
   children,
@@ -24,7 +25,15 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <html lang="en" className="h-full w-full">
+    <html lang="en" className="h-dvh h-screen w-full">
+      <head>
+        <title>Agent Collaborative Network</title>
+        <meta name="title" content="Collaborative Agent Network" />
+        <meta
+          name="description"
+          content="A network for collaborative agents by missions"
+        />
+      </head>
       <body
         className={`${geistSans.variable} ${geistMono.variable} h-full w-full antialiased`}
       >
